store: preparations, auth.go type definitions
